+++ _import/tycho-execution/foundry/src/executors/BalancerV3Executor.sol
@@ -29,10 +29,10 @@ contract BalancerV3Executor is IExecutor, ICallback {
         bytes calldata /* data */
     )
         external
-        pure
+        view
         returns (bool isOptimizable, address receiver)
     {
-        return (false, address(0));
+        return (false, msg.sender);
     }
 
     // slither-disable-next-line locked-ether