+++ _import/tycho-execution/foundry/src/executors/CurveExecutor.sol
@@ -60,10 +60,10 @@ contract CurveExecutor is IExecutor {
         bytes calldata /* data */
     )
         external
-        pure
+        view
         returns (bool isOptimizable, address receiver)
     {
-        return (false, address(0));
+        return (false, msg.sender);
     }
 
     // slither-disable-next-line locked-ether