+++ _import/tycho-execution/foundry/src/executors/EkuboExecutor.sol
@@ -53,10 +53,10 @@ contract EkuboExecutor is IExecutor, ILocker, IPayer, ICallback {
         bytes calldata /* data */
     )
         external
-        pure
+        view
         returns (bool isOptimizable, address receiver)
     {
-        return (false, address(0));
+        return (false, msg.sender);
     }
 
     function swap(uint256 amountIn, bytes calldata data, address receiver)