+++ _import/tycho-execution/foundry/src/executors/HashflowExecutor.sol
@@ -52,10 +52,10 @@ contract HashflowExecutor is IExecutor {
         bytes calldata /* data */
     )
         external
-        pure
+        view
         returns (bool isOptimizable, address receiver)
     {
-        return (false, address(0));
+        return (false, msg.sender);
     }
 
     function swap(uint256 amountIn, bytes calldata data, address receiver)