+++ _import/tycho-execution/foundry/src/executors/SlipstreamsExecutor.sol
@@ -53,10 +53,10 @@ contract SlipstreamsExecutor is IExecutor, ICallback {
         bytes calldata /* data */
     )
         external
-        pure
+        view
         returns (bool isOptimizable, address receiver)
     {
-        return (false, address(0));
+        return (false, msg.sender);
     }
 
     // slither-disable-next-line locked-ether