+++ _import/tycho-execution/foundry/src/executors/UniswapV3Executor.sol
@@ -45,10 +45,10 @@ contract UniswapV3Executor is IExecutor, ICallback {
         bytes calldata /* data */
     )
         external
-        pure
+        view
         returns (bool isOptimizable, address receiver)
     {
-        return (false, address(0));
+        return (false, msg.sender);
     }
 
     // slither-disable-next-line locked-ether